Unified command traits, changed take error signature, reduced generics in API
