+++ src/lib.rs
@@ -15,10 +15,12 @@ pub use crate::console::{
     ConsoleCommand, ConsoleCommandEntered, ConsoleConfiguration, ConsoleOpen, PrintConsoleLine,
     ToggleConsoleKey,
 };
-pub use crate::value::{FromValue, FromValueError, ValueType};
+pub use crate::error::FromValueError;
+pub use crate::value::{FromValue, ValueType};
 
 mod commands;
 mod console;
+mod error;
 mod macros;
 mod value;
 
@@ -32,9 +34,9 @@ impl Plugin for ConsolePlugin {
             .init_resource::<ConsoleOpen>()
             .add_event::<ConsoleCommandEntered>()
             .add_event::<PrintConsoleLine>()
-            .add_console_command::<ClearCommand, _, _>(clear_command)
-            .add_console_command::<ExitCommand, _, _>(exit_command)
-            .add_console_command::<HelpCommand, _, _>(help_command)
+            .add_console_command::<ClearCommand, _>(clear_command)
+            .add_console_command::<ExitCommand, _>(exit_command)
+            .add_console_command::<HelpCommand, _>(help_command)
             .add_system(console_ui.at_end())
             .add_system(receive_console_line);
